requirements.txt
amqp==5.3.1
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.10
attrs==25.3.0
bandit==1.8.5
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cssselect2==0.8.0
dill==0.4.0
Django==5.2.3
django-celery-beat==2.5.0
django-crispy-forms==2.4
django-jazzmin==3.0.1
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.2
django-simple-history==3.8.0
et_xmlfile==2.0.0
flake8==7.2.0
Flask==3.1.1
fonttools==4.58.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
httplib2==0.22.0
ics==0.7.2
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.16.1
mypy_extensions==1.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pbr==6.1.1
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyflakes==3.3.2
Pygments==2.19.1
pylint==3.3.7
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stevedore==5.4.1
TatSu==5.13.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
XlsxWriter==3.2.3
zopfli==0.2.3.post1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
django-cors-headers==4.6.0
scikit-learn==1.5.1
scipy==1.14.1
django-otp==1.5.4
qrcode[pil]==8.0.1
django-axes==6.5.1
django-ratelimit==4.1.0
django-modeltranslation==0.19.8
